requirements.txt
pystray>=0.19.5
customtkinter>=5.2.2
mss>=9.0.1
requests>=2.31.0
Pillow>=10.0.0
psutil>=5.9.8
pyinstaller>=6.0.0
